修bug

+++ src/main/java/com/p1nero/invincible/conditions/HealthCondition.java
@@ -10,7 +10,7 @@ public class HealthCondition implements Condition<ServerPlayerPatch>{
     private final boolean isTarget, isLarger;
     private final float ratio;
 
-    public HealthCondition(boolean isTarget, int ratio, boolean isLarger) {
+    public HealthCondition(boolean isTarget, float ratio, boolean isLarger) {
         this.isTarget = isTarget;
         this.ratio = ratio;
         this.isLarger = isLarger;
